adding search task

+++ AiDevReloaded.Core/OpenAI/Embedding/EmbeddingServiceExtensions.cs
@@ -2,7 +2,7 @@
 
 public static class EmbeddingServiceExtensions
 {
-    public static async Task<double[]> GetEmbedding(this EmbeddingService service, string input, CancellationToken cancellationToken = default)
+    public static async Task<float[]> GetEmbedding(this EmbeddingService service, string input, CancellationToken cancellationToken = default)
     {
         var response = await service.GetEmbeddingInfo(input, cancellationToken);
         return response.Data.First().Embedding.ToArray();
