+++ AiDevReloaded.Tests/C02L03_embedding/Test.cs
@@ -9,7 +9,7 @@ public sealed class Test
     [Fact]
     public async Task Should_Pass_Embedding_Task()
     {
-        var taskSession = new TaskSession<DefaultResponse, double[]>(Secrets.AiDevTaskApiKey, "embedding");
+        var taskSession = new TaskSession<DefaultResponse, float[]>(Secrets.AiDevTaskApiKey, "embedding");
         var taskContent = await taskSession.GetTaskContent();
         var sentence = "Hawaiian pizza";
         var embeddingService = new EmbeddingService(Secrets.OpenAiApiKey);
